# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Spotplay
  description: 'Es una plataforma de musica, puedes estar al tanto de las nuevas canciones de tus artistas favoritos, Crea una cuenta con nosotros para publicar tus canciones o para escuchar las canciones favoritas de tus artistas'
  contact:
    email: axelgolin4@gmail.com
  license:
    name: Apache 2.0
    url: http://apache.com
  version: 1.0.0
servers:
- url: http://localhost:4000/api/v1
...
components:
  schemas:
    Song:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        uri:
          type: string
        duration:
          type: string
        imagen: 
          type: string
        idAlbum:
          type: integer
          format: int64
        idArtist:
          type: integer
          format: int32
        idGenre:
          type: integer
          format: int32
...
paths:
  /song:
    post:
      tags:
      - Song
      summary: Crea una cancion nueva.
      operationId: createSong
      description: Se crea una nueva cancion.
      requestBody:
        description: Ingrese los datos de la cancion a registrar.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
        required: true
      responses: 
        200: 
          description: Registro exitoso
          content:
              application/json:
                schema: 
                  $ref: '#/components/schemas/Song'
        400:
          description: bad request
          content: {}
        404:
          description:  Not found
          content: {}
    x-codegen-request-body-name: body
    get:
      tags:
      - Song
      summary: Obtiene todas las canciones publicadas
      operationId: getAllSongs
      description: Se obtiene todas las canciones publicadas
      responses:
        200:
          description: Obtencion de las canciones con exito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        400:
          description: bad request
          content: {}
        404:
          description:  Not found
          content: {}
    delete:
      tags:
      - Song
      summary: Eliminar una cancion.
      operationId: deleteSong
      description: Se elimina una cancion.
      requestBody:
        description: Ingrese la cancion que se eliminara.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
        required: true
      responses: 
        200: 
          description: Registro exitoso
          content:
              application/json:
                schema: 
                  $ref: '#/components/schemas/Song'
        400:
          description: bad request
          content: {}
        404:
          description:  Not found
          content: {}
    x-codegen-request-body-name: body
    put:
      tags:
      - Song
      summary: Editar una cancion.
      operationId: updateSong
      description: Se editara una cancion.
      requestBody:
        description: Ingrese la cancion que se editara.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
        required: true
      responses: 
        200: 
          description: Registro exitoso
          content:
              application/json:
                schema: 
                  $ref: '#/components/schemas/Song'
        400:
          description: bad request
          content: {}
        404:
          description:  Not found
          content: {}
    x-codegen-request-body-name: body
...